requirements.txt
streamlit
pandas
sqlalchemy
pymongo
mysql-connector-python
psycopg2-binary
matplotlib
seaborn
gunicorn
